requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
llama-cpp-python
streamlit
protobuf==3.20.0
google-genai
pypdf
